requirements.txt
kivy>=2.1.0
icalendar>=4.0.7
requests>=2.27.1
reportlab>=3.6.9
python-dateutil>=2.8.2
pillow>=9.0.0
```